# src/openapi-spec.json
{
  "components": {
    "schemas": {
      "AccountUser": { "properties": {}, "type": "object" },
      "CreateAccountUserDto": { "properties": {}, "type": "object" },
      "CreateCoffeeDto": { "properties": {}, "type": "object" },
      "CreatePostDto": {
        "properties": {
          "content": { "type": "string" },
          "title": { "description": "The title of the post.", "type": "string" }
        },
        "required": ["content", "title"],
        "type": "object"
      },
      "CreateTagDto": { "properties": {}, "type": "object" },
      "ErrorResponse": {
        "properties": {
          "path": { "example": "/request/path", "type": "string" },
          "statusCode": { "example": 400, "type": "number" },
          "timestamp": {
            "example": "2023-12-29T16:52:26.720Z",
            "type": "string"
          }
        },
        "required": ["path", "statusCode", "timestamp"],
        "type": "object"
      },
      "ListPostResponse": {
        "properties": {
          "pagination": { "$ref": "#/components/schemas/PaginationDto" },
          "posts": {
            "items": { "$ref": "#/components/schemas/PostResponse" },
            "type": "array"
          },
          "totalCount": { "type": "number" }
        },
        "required": ["pagination", "posts", "totalCount"],
        "type": "object"
      },
      "PaginationDto": {
        "properties": {
          "page": { "type": "number" },
          "perPage": { "type": "number" }
        },
        "required": ["page", "perPage"],
        "type": "object"
      },
      "PaymentWebhookResponse": {
        "properties": { "success": { "type": "boolean" } },
        "required": ["success"],
        "type": "object"
      },
      "PostResponse": {
        "properties": {
          "content": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "removed": { "type": "boolean" },
          "title": { "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "content",
          "createdAt",
          "id",
          "removed",
          "title",
          "updatedAt"
        ],
        "type": "object"
      },
      "RemovePostsDto": { "properties": {}, "type": "object" },
      "UpdateAccountUserDto": { "properties": {}, "type": "object" },
      "UpdateCoffeeDto": { "properties": {}, "type": "object" },
      "UpdateMeDto": {
        "properties": { "name": { "example": "bob", "type": "string" } },
        "required": ["name"],
        "type": "object"
      },
      "UpdatePostDto": {
        "properties": {
          "content": {
            "description": "new content of the post",
            "type": "string"
          },
          "title": { "description": "new title of the post", "type": "string" }
        },
        "type": "object"
      },
      "UpdateTagDto": { "properties": {}, "type": "object" },
      "UserResponse": {
        "properties": {
          "id": {
            "example": "886e8222-9aff-4fd3-87dd-02344242a17c",
            "type": "string"
          },
          "name": { "example": "alice", "type": "string" }
        },
        "required": ["id", "name"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "DOCUMENT_JWT_AUTH_NAME": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "user and post api description",
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "title": "nestjs template",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/account-users": {
      "get": {
        "operationId": "AccountUsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/AccountUser" },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "get posts",
        "tags": ["account-users"]
      },
      "post": {
        "operationId": "AccountUsersController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAccountUserDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountUser" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "create post",
        "tags": ["account-users"]
      }
    },
    "/account-users/{id}": {
      "delete": {
        "operationId": "AccountUsersController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountUser" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "delete post",
        "tags": ["account-users"]
      },
      "get": {
        "operationId": "AccountUsersController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountUser" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "get post",
        "tags": ["account-users"]
      },
      "patch": {
        "operationId": "AccountUsersController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAccountUserDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountUser" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "update post",
        "tags": ["account-users"]
      }
    },
    "/coffees": {
      "get": {
        "operationId": "CoffeesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "get coffees",
        "tags": ["posts"]
      },
      "post": {
        "operationId": "CoffeesController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCoffeeDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "create coffee",
        "tags": ["posts"]
      }
    },
    "/coffees/{id}": {
      "delete": {
        "operationId": "CoffeesController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "delete coffee",
        "tags": ["posts"]
      },
      "get": {
        "operationId": "CoffeesController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "get coffee",
        "tags": ["posts"]
      },
      "patch": {
        "operationId": "CoffeesController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCoffeeDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "update coffee",
        "tags": ["posts"]
      }
    },
    "/fibonacci": {
      "get": {
        "operationId": "FibonacciController_fibonacci",
        "parameters": [
          {
            "in": "query",
            "name": "n",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "number" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "get coffee",
        "tags": ["fibonacci"]
      }
    },
    "/i18n/greeting": {
      "get": {
        "operationId": "I18nController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "receive greeting, usage\n      curl -H 'accept-language: pl' 'localhost:3000/i18n/greeting'\n      curl -H 'accept-language: en' 'localhost:3000/i18n/greeting'",
        "tags": ["i18n"]
      }
    },
    "/payments-webhook": {
      "get": {
        "operationId": "PaymentsWebhookController_webhook",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "webhook for payment"
      }
    },
    "/posts": {
      "delete": {
        "operationId": "PostController_removePosts",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RemovePostsDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "delete requested user's posts",
        "tags": ["posts"]
      },
      "get": {
        "operationId": "PostController_listPost",
        "parameters": [
          {
            "example": 20,
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "default": 20, "type": "number" }
          },
          {
            "example": 0,
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": { "default": 0, "type": "number" }
          },
          {
            "example": "asc",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": ["asc", "desc"],
              "type": "string"
            }
          },
          {
            "example": "createdAt",
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "default": "createdAt",
              "enum": ["createdAt", "updatedAt"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListPostResponse" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "get requested user's posts",
        "tags": ["posts"]
      },
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePostDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "create requested user's post",
        "tags": ["posts"]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostController_getPost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "get requested user's post",
        "tags": ["posts"]
      },
      "put": {
        "operationId": "PostController_putPost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePostDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "update requested user's post",
        "tags": ["posts"]
      }
    },
    "/tags": {
      "get": {
        "operationId": "TagsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "TBW"
      },
      "post": {
        "operationId": "TagsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTagDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "TBW"
      }
    },
    "/tags/{id}": {
      "delete": {
        "operationId": "TagsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "TBW"
      },
      "get": {
        "operationId": "TagsController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "TBW"
      },
      "patch": {
        "operationId": "TagsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTagDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "string" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "summary": "TBW"
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "get requested user's profile",
        "tags": ["users"]
      },
      "patch": {
        "operationId": "UserController_updateMe",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMeDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "DOCUMENT_JWT_AUTH_NAME": [] }],
        "summary": "update requested user's profile",
        "tags": ["users"]
      }
    }
  },
  "servers": [],
  "tags": []
}